moved some things around

+++ imc-efs-automation-service/imc-efs-automation-service-core/src/main/java/com/imc/efs/automation/dao/CheckLimitDAO.java
@@ -1,6 +1,8 @@
 package com.imc.efs.automation.dao;
 
+import java.math.BigDecimal;
+
 public interface CheckLimitDAO {
 
-	double getEfsCheckLimitByUser(String requester, int requestTypeId);
+	BigDecimal getEfsCheckLimitByUser(String requester, int requestTypeId);
 }
