+++ imc-efs-automation-service/imc-efs-automation-service-core/src/main/java/com/imc/efs/automation/dao/EfsDAO.java
@@ -8,7 +8,7 @@ import com.imc.efs.automation.data.EfsMoneyCode;
 public interface EfsDAO {
 
 	EfsMoneyCode getMoneyCode(BigDecimal efsAmount, String issueTo, String description, String company);
-	EfsMoneyCode GetMoneyCode_TEST(double efsAmount, String issueTo, String description, String company);
+	EfsMoneyCode GetMoneyCode_TEST(BigDecimal efsAmount, String issueTo, String description, String company);
 	MoneyCodeDetailsDTO GetMoneyCodeDetails(String company, int referenceNumber);
 	MoneyCodeDetailsDTO GetMoneyCodeDetails_TEST(String company, int referenceNumber);
 }
